! 2012-08-09 - Chris Jones  - Explicitly caste variables to float when filling the training tuple  - Improvements to the training job scripts
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@143931 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 79f4b0fc7dde8ca75ce2c313bc37e634bf8fef12

+++ Rec/Rec/ChargedProtoANNPID/src/ChargedProtoANNPIDTupleTool.cpp
@@ -121,7 +121,7 @@ StatusCode ChargedProtoANNPIDTupleTool::fill( const LHCb::ProtoParticle * proto,
   for ( ; inputS != m_variablesS.end() && inputI != m_variablesI.end(); ++inputS, ++inputI )
   {
     // get the variable and fill ntuple
-    sc = sc && tuple->column( *inputS, getInput(proto,*inputI) );
+    sc = sc && tuple->column( *inputS, (float)getInput(proto,*inputI) );
   }
 
   // MC variables
